+++ .github/cicd/scripts/check-release-notes.ts
@@ -1,6 +1,6 @@
 import { existsSync } from "@std/fs/exists";
-import { Utils } from "../../../core/Utils.ts";
-import getEnvVar from "../core/GetEnvVar.ts";
+import { Utils } from "@core/Utils.ts";
+import getEnvVar from "@cicd/core/GetEnvVar.ts";
 
 const scriptFileName = new URL(import.meta.url).pathname.split("/").pop();
