+++ .github/cicd/scripts/get-version.ts
@@ -1,4 +1,4 @@
-import getEnvVar from "../core/GetEnvVar.ts";
+import getEnvVar from "@cicd/core/GetEnvVar.ts";
 import DenoConfig from "../../../deno.json" with { type: "json" };
 
 const scriptFileName = new URL(import.meta.url).pathname.split("/").pop();
