+++ .github/cicd/scripts/github-release-does-not-exist.ts
@@ -1,6 +1,6 @@
 import { ReleaseClient } from "@kd/clients";
-import getEnvVar from "../core/GetEnvVar.ts";
-import { validateOrgExists, validateRepoExists, validateUserExists } from "../core/Validators.ts";
+import getEnvVar from "@cicd/core/GetEnvVar.ts";
+import { validateOrgExists, validateRepoExists, validateUserExists } from "@cicd/core/Validators.ts";
 
 const scriptFileName = new URL(import.meta.url).pathname.split("/").pop();
