+++ .github/cicd/scripts/milestone-exists.ts
@@ -1,5 +1,5 @@
-import getEnvVar from "../core/GetEnvVar.ts";
-import { validateMilestoneExists, validateOrgExists, validateRepoExists, validateUserExists } from "../core/Validators.ts";
+import getEnvVar from "@cicd/core/GetEnvVar.ts";
+import { validateMilestoneExists, validateOrgExists, validateRepoExists, validateUserExists } from "@cicd/core/Validators.ts";
 
 const scriptFileName = new URL(import.meta.url).pathname.split("/").pop();
