+++ .github/cicd/scripts/validate-version.ts
@@ -1,5 +1,5 @@
-import { Utils } from "../../../core/Utils.ts";
-import getEnvVar from "../core/GetEnvVar.ts";
+import { Utils } from "@core/Utils.ts";
+import getEnvVar from "@cicd/core/GetEnvVar.ts";
 
 const scriptFileName = new URL(import.meta.url).pathname.split("/").pop();
