+++ GitHubClients/LabelClient.ts
@@ -1,10 +1,10 @@
-import { AuthError } from "../deps.ts";
-import { Utils } from "../deps.ts";
-import { GitHubHttpStatusCodes } from "../core/Enums.ts";
-import { GitHubClient } from "../deps.ts";
-import { Guard } from "../core/Guard.ts";
-import { LabelError } from "../deps.ts";
-import type { LabelModel } from "../deps.ts";
+import { Utils } from "@core/Utils.ts";
+import { GitHubHttpStatusCodes } from "@core/Enums.ts";
+import { GitHubClient } from "@gh-clients/GitHubClient.ts";
+import { Guard } from "@core/Guard.ts";
+import { AuthError } from "@gh-errors/AuthError.ts";
+import { LabelError } from "@gh-errors/LabelError.ts";
+import type { LabelModel } from "@gh-rest-models/LabelModel.ts";
 
 /**
  * Provides a client for interacting with labels.
