+++ PackageClients/mod.ts
@@ -1,3 +1,3 @@
 // NOTE: Do not refactor the import paths below to use the import maps.  Leave them as relative paths.
 export { NuGetClient } from "./NuGetClient.ts";
-export { NuGetError } from "../PackageClients/Errors/NuGetError.ts";
+export { NuGetError } from "@pkg-errors/NuGetError.ts";
