+++ core/GraphQl/Queries/GetBranchesQuery.ts
@@ -1,4 +1,4 @@
-import { Utils } from "../../../deps.ts";
+import { Utils } from "@core/Utils.ts";
 
 /**
  * Get a list of branches for a GitHub repository.
