+++ core/IssueOrPRRequestData.ts
@@ -1,4 +1,4 @@
-import type { IssueState, StateReason } from "./Enums.ts";
+import type { IssueState, StateReason } from "@core/Enums.ts";
 
 /**
  * Represents the body of an HTTP GitHub API issue request.
